+++ src/main/java/fr/raksrinana/fallingtree/tree/TreeHandler.java
@@ -1,6 +1,8 @@
 package fr.raksrinana.fallingtree.tree;
 
 import fr.raksrinana.fallingtree.FallingTree;
+import fr.raksrinana.fallingtree.utils.ToAnalyzePos;
+import fr.raksrinana.fallingtree.utils.TreePart;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.entity.player.PlayerEntity;
@@ -12,24 +14,23 @@ import net.minecraft.util.math.Direction;
 import net.minecraft.world.World;
 import java.util.*;
 import java.util.stream.Collectors;
-import static fr.raksrinana.fallingtree.FallingTreeUtils.isLeafBlock;
-import static fr.raksrinana.fallingtree.FallingTreeUtils.isTreeBlock;
+import static fr.raksrinana.fallingtree.FallingTreeUtils.*;
 
 public class TreeHandler{
-	public static Optional<Tree> getTree(World world, BlockPos blockPos){
-		Block logBlock = world.getBlockState(blockPos).getBlock();
-		if(!isTreeBlock(logBlock)){
+	public static Optional<Tree> getTree(World world, BlockPos originPos){
+		Block originBlock = world.getBlockState(originPos).getBlock();
+		if(!isLogBlock(originBlock)){
 			return Optional.empty();
 		}
-		Queue<BlockPos> toAnalyzePos = new LinkedList<>();
-		Set<BlockPos> analyzedPos = new HashSet<>();
-		Tree tree = new Tree(world, blockPos);
-		toAnalyzePos.add(blockPos);
+		Queue<ToAnalyzePos> toAnalyzePos = new PriorityQueue<>();
+		Set<ToAnalyzePos> analyzedPos = new HashSet<>();
+		Tree tree = new Tree(world, originPos);
+		toAnalyzePos.add(new ToAnalyzePos(originPos, originBlock, originPos, originBlock, TreePart.LOG));
 		while(!toAnalyzePos.isEmpty()){
-			BlockPos analyzingPos = toAnalyzePos.remove();
-			tree.addLog(analyzingPos);
+			ToAnalyzePos analyzingPos = toAnalyzePos.remove();
+			tree.addPart(analyzingPos.getTreePart(), analyzingPos.getCheckPos());
 			analyzedPos.add(analyzingPos);
-			Collection<BlockPos> nearbyPos = neighborLogs(world, logBlock, analyzingPos, analyzedPos);
+			Collection<ToAnalyzePos> nearbyPos = neighborLogs(world, originPos, originBlock, analyzingPos, analyzedPos);
 			nearbyPos.removeAll(analyzedPos);
 			toAnalyzePos.addAll(nearbyPos.stream().filter(pos -> !toAnalyzePos.contains(pos)).collect(Collectors.toList()));
 		}
@@ -46,15 +47,18 @@ public class TreeHandler{
 		return Optional.of(tree);
 	}
 	
-	private static Collection<BlockPos> neighborLogs(World world, Block logBlock, BlockPos blockPos, Collection<BlockPos> analyzedPos){
-		List<BlockPos> neighborLogs = new LinkedList<>();
+	private static Collection<ToAnalyzePos> neighborLogs(World world, BlockPos originPos, Block originBlock, ToAnalyzePos parent, Collection<ToAnalyzePos> analyzedPos){
+		final BlockPos blockPos = parent.getCheckPos();
+		List<ToAnalyzePos> neighborLogs = new LinkedList<>();
 		final BlockPos.Mutable checkPos = new BlockPos.Mutable();
 		for(int x = -1; x <= 1; x++){
 			for(int z = -1; z <= 1; z++){
 				for(int y = -1; y <= 1; y++){
 					checkPos.set(blockPos.getX() + x, blockPos.getY() + y, blockPos.getZ() + z);
-					if(!analyzedPos.contains(checkPos) && isSameTree(world, checkPos, logBlock)){
-						neighborLogs.add(checkPos.toImmutable());
+					Block checkBlock = world.getBlockState(checkPos).getBlock();
+					ToAnalyzePos analyzed = new ToAnalyzePos(blockPos, parent.getCheckBlock(), checkPos.toImmutable(), checkBlock, getTreePart(checkBlock));
+					if(!analyzedPos.contains(analyzed) && shouldIncludeInChain(originPos, originBlock, parent, analyzed)){
+						neighborLogs.add(analyzed);
 					}
 				}
 			}
@@ -63,20 +67,47 @@ public class TreeHandler{
 		return neighborLogs;
 	}
 	
+	private static TreePart getTreePart(Block checkBlock){
+		if(isLogBlock(checkBlock)){
+			return TreePart.LOG;
+		}
+		if(isNetherWartOrShroomlight(checkBlock)){
+			return TreePart.WART;
+		}
+		return TreePart.OTHER;
+	}
+	
 	private static long getLeavesAround(World world, BlockPos blockPos){
 		return Arrays.stream(Direction.values())
 				.map(blockPos::offset)
-				.filter(testPos -> isLeafBlock(world.getBlockState(testPos).getBlock()))
+				.filter(testPos -> {
+					Block block = world.getBlockState(testPos).getBlock();
+					return isLeafBlock(block) || isNetherWartOrShroomlight(block);
+				})
 				.count();
 	}
 	
-	private static boolean isSameTree(World world, BlockPos checkBlockPos, Block parentLogBlock){
-		Block checkBlock = world.getBlockState(checkBlockPos).getBlock();
+	private static boolean shouldIncludeInChain(BlockPos originPos, Block originBlock, ToAnalyzePos parent, ToAnalyzePos check){
+		if(parent.getTreePart() == TreePart.LOG && isSameTree(originBlock, check)){
+			return true;
+		}
+		if(FallingTree.config.trees.isBreakNetherTreeWarts()){
+			if(check.getTreePart() == TreePart.WART){
+				BlockPos checkBlockPos = check.getCheckPos();
+				int dx = Math.abs(originPos.getX() - checkBlockPos.getX());
+				int dz = Math.abs(originPos.getZ() - checkBlockPos.getZ());
+				return dx <= 4 && dz <= 4;
+			}
+		}
+		return false;
+	}
+	
+	private static boolean isSameTree(Block parentLogBlock, ToAnalyzePos check){
 		if(FallingTree.config.getTreesConfiguration().isAllowMixedLogs()){
-			return isTreeBlock(checkBlock);
+			return check.getTreePart() == TreePart.LOG;
 		}
 		else{
-			return checkBlock.equals(parentLogBlock);
+			return check.getCheckBlock().equals(parentLogBlock);
 		}
 	}
 	
@@ -105,6 +136,11 @@ public class TreeHandler{
 			tool.damage(toolDamage, player, (entity) -> {});
 		}
 		if(isTreeFullyBroken){
+			tree.getWarts().forEach(wartPos -> {
+				final BlockState wartState = world.getBlockState(wartPos);
+				wartState.getBlock().afterBreak(world, player, wartPos, wartState, world.getBlockEntity(wartPos), tool);
+				world.removeBlock(wartPos, false);
+			});
 			final int radius = FallingTree.config.getTreesConfiguration().getLeavesBreakingForceRadius();
 			if(radius > 0){
 				tree.getLogs().stream().max(Comparator.comparingInt(BlockPos::getY)).ifPresent(topLog -> {
@@ -139,7 +175,7 @@ public class TreeHandler{
 				return false;
 			}
 		}
-		tree.getTopMostFurthestLog().ifPresent(logBlock -> {
+		tree.getTopMostFurthestPart().ifPresent(logBlock -> {
 			final BlockState logState = world.getBlockState(logBlock);
 			logState.getBlock().afterBreak(world, player, tree.getHitPos(), logState, world.getBlockEntity(logBlock), tool);
 			world.removeBlock(logBlock, false);